pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grant_finder"
version = "0.1.0"
authors = [
    { name = "Grant Finder Team" },
]
description = "LangGraph-based grant opportunity finder"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core LangGraph/LangChain dependencies
    "langgraph>=0.0.15",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.16",
    
    # Vector store and embeddings
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.2",
    
    # Document processing
    "unstructured>=0.11.0",
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "pdfminer.six>=20221105",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    
    # Web tools
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "selenium>=4.15.2",
    "webdriver-manager>=4.0.1",
    
    # Utility libraries
    "pydantic>=2.5.2",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "tqdm>=4.66.1",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3.post1",
    
    # Logging and monitoring
    "rich>=13.7.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "ruff>=0.1.6",
]

test = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "aiohttp>=3.9.1",
    "respx>=0.20.2",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "B",  # flake8-bugbear
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=grant_finder --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["grant_finder"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/grant_finder"
"Bug Tracker" = "https://github.com/yourusername/grant_finder/issues"